Крестики нолики Две Миланы в домике Укатились даже шарики за ролики Улы-улыбаемся, разгоняя тучки Люди обзываются что мы две почемучки Источник text-pesni.com  Отпути, не путю Но кому я говорю Не путю ведь я люблю Свою лучшую подругу Отпути, не путю Но кому я говорю Не путю ведь я люблю Навсегда мы друг для друга  ЛП  Эники беники В ютубе и на телеке Мы тик ток доводим топы до истерики Сразу все влюбляются Когда хочу на ручки Люди обзываются что мы две липучки  Отпути, не путю Но кому я говорю Не путю ведь я люблю Свою лучшую подругу Отпути, не путю Но кому я говорю Не путю ведь я люблю Навсегда мы друг для друга ЛП  У нас с тобой свой Дисней Нам не надо кукол Повсюду за мной везде Моя лучшая подруга ЛП
Крестики нолики
Две Миланы в домике
Укатились даже шарики за ролики
Улы-улыбаемся, разгоняя тучки
Люди обзываются что мы две почемучки
Источник text-pesni.com

Отпути, не путю
Но кому я говорю
Не путю ведь я люблю
Свою лучшую подругу
Отпути, не путю
Но кому я говорю
Не путю ведь я люблю
Навсегда мы друг для друга

ЛП

Эники беники
В ютубе и на телеке
Мы тик ток доводим топы до истерики
Сразу все влюбляются
Когда хочу на ручки
Люди обзываются что мы две липучки

Отпути, не путю
Но кому я говорю
Не путю ведь я люблю
Свою лучшую подругу
Отпути, не путю
Но кому я говорю
Не путю ведь я люблю
Навсегда мы друг для друга
ЛП

У нас с тобой свой Дисней
Нам не надо кукол
Повсюду за мной везде
Моя лучшая подруга
ЛП

+++ Assets/скрипты/враги/Enemy.cs
@@ -84,10 +84,13 @@ public abstract class Enemy : MonoBehaviour
 
         if (health <= 0)
         {
-            Manager.Instance.totalKills++;
-            Manager.Instance.levelKills++;
+            if (maxHealth > 1)
+            {                
+                Manager.Instance.totalKills++;
+                Manager.Instance.levelKills++;
+                Manager.Instance.Kill(damage, target.Value, transform, crit, source);
+            }
 
-            Manager.Instance.Kill(damage, target.Value, transform, crit, source);
             gameObject.tag = "Untagged";
             if (maxHealth > 1) Destroy(healthBar.gameObject);
             for (int i = 0; i < 10; i++) Blood(target.Value);
